+++ test/unit/Jamfile
@@ -34,7 +34,6 @@ run
         test/protocol/protocol_types.cpp
         test/protocol/text_protocol.cpp
         test/protocol/binary_protocol.cpp
-        test/protocol/query_with_params.cpp
         test/protocol/serialization.cpp
         test/protocol/deserialization.cpp
